+++ src/main/kotlin/ru/my/test/entity/BaseEntity.kt
@@ -9,5 +9,5 @@ import javax.persistence.MappedSuperclass;
 open class BaseEntity(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Int
+    var id: Long
 )
