+++ src/main/kotlin/ru/my/test/model/Author.kt
@@ -4,20 +4,20 @@ import org.hibernate.validator.constraints.UniqueElements
 import javax.validation.constraints.NotNull
 
 class AuthorView(
-    val id: Int,
+    val id: Long,
     val name: String,
-    val bookIds: List<Int> = mutableListOf(),
+    val bookIds: List<Long> = mutableListOf(),
 )
 
 class AuthorAddRequest(
     @field:NotNull
     val name: String,
     @field:UniqueElements
-    val bookIds: List<Int> = mutableListOf(),
+    val bookIds: List<Long> = mutableListOf(),
 )
 
 class AuthorEditRequest(
     val name: String,
     @field:UniqueElements
-    val bookIds: List<Int> = mutableListOf(),
+    val bookIds: List<Long> = mutableListOf(),
 )
