+++ src/main/kotlin/ru/my/test/model/Book.kt
@@ -8,9 +8,9 @@ import javax.validation.constraints.NotNull
 
 @JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator::class, property = "id")
 class BookView(
-    val id: Int,
+    val id: Long,
     val name: String,
-    val authorIds: List<Int> = mutableListOf(),
+    val authorIds: List<Long> = mutableListOf(),
     val reviews: List<ReviewView> = mutableListOf(),
 )
 
@@ -19,12 +19,12 @@ class BookAddRequest(
     @field:UniqueBookName
     val name: String,
     @field:UniqueElements
-    val authorIds: List<Int> = mutableListOf(),
+    val authorIds: List<Long> = mutableListOf(),
 )
 
 class BookEditRequest(
     @field:UniqueBookName
     val name: String,
     @field:UniqueElements
-    val authorIds: List<Int> = mutableListOf(),
+    val authorIds: List<Long> = mutableListOf(),
 )
