+++ src/main/kotlin/ru/my/test/service/ReviewRepository.kt
@@ -5,11 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository
 import ru.my.test.entity.Book
 import ru.my.test.entity.Review
 
-interface ReviewRepository : JpaRepository<Review, Int> {
-    fun findByBookId(bookId: Int): List<Review>
+interface ReviewRepository : JpaRepository<Review, Long> {
+    fun findByBookId(bookId: Long): List<Review>
 }
 
-fun ReviewRepository.findOrException(id: Int): Review {
+fun ReviewRepository.findOrException(id: Long): Review {
     return this.findById(id).orElseThrow {
         NotFoundException("Не удалось найти отзыв с id: $id")
     }
